A patch from Deepak for "Incorrect use of if-has-permission tag" https://issues.apache.org/jira/browse/OFBIZ-5920
Deepak: here is the patch for this issue, found some more places where if-has-permission tag used in different way, will create a new ticket to fix the if-has-permission tags and will used the consistent pattern.

jleroux: I wonder if this commit does not fix some other already reported issues, anyway it's good!

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/humanres@1647266 13f79535-47bb-0310-9956-ffa450edef68