Improved: Document the rounting number fields in entities (OFBIZ-7725)
Initially Swapnil Shah proposed to add a branchCode field in EftAccount and 
Check entities. Pierre mentioned that branchCode is actually the same than the
existing rountingNumber field there.

This adds a description for the rountingNumber field (also in PayrollPreference 
in humares) with a link to  https://en.wikipedia.org/wiki/Bank_code
where this field is explained for many countries.

As suggested by Paul Foxworthy I added the Bank-State-Branch label for 
Australia. You might want to check the translation for you country or add it

Finally I removed the FormFieldTitle_routingNumber in in humanres, it's not used
anywhere and at least the French translation was wrong. Better to stick to 
AccountingRoutingNumber

Thanks: Swapnil, Pierre and Paul for the helpful discussion

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/humanres@1770345 13f79535-47bb-0310-9956-ffa450edef68